+++ ranking-form-frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from './logo.svg';
+import logo from 'assets/logo.svg';
 import './App.css';
 
 function App() {
