+++ ranking-form-frontend/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from 'lib/reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
